Conflicts:
	client_src/add_vrpn_cookie.C
	vrpn_3DConnexion.C
	vrpn_Connection.C

clang-tidy and Russ solved printf mismatches differently.

+++ client_src/add_vrpn_cookie.C
@@ -71,8 +71,8 @@ int main (int argc, char ** argv) {
   if (retval != vrpn_cookie_size()) {
     fprintf(stderr, "vrpn_Connection::close_log:  "
                     "Couldn't write magic cookie to log file "
-                    "(got %d, expected %lu).\n",
-            retval, vrpn_cookie_size());
+                    "(got %d, expected %d).\n",
+            retval, static_cast<int>(vrpn_cookie_size()));
     goto CLEANUP;
   }
